refactor: extract reusable logic into custom hooks
