+++ src/features/tracks/components/track-form.tsx
@@ -1,8 +1,5 @@
-import { zodResolver } from "@hookform/resolvers/zod";
 import { AlertCircle, CheckCircle, FileImage, Loader2, X } from "lucide-react";
 import Image from "next/image";
-import { useCallback, useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
 
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
@@ -18,8 +15,8 @@ import {
 import { Input } from "@/components/ui/input";
 import { ScrollArea } from "@/components/ui/scroll-area";
 
-import { useGenres } from "../hooks/use-genres";
-import { TrackFormData, TrackFormSchema } from "../schemas/schemas";
+import { useTrackForm } from "../hooks/use-track-form";
+import { TrackFormData } from "../schemas/schemas";
 
 interface TrackFormProps {
   initialData?: Partial<TrackFormData>;
@@ -34,134 +31,30 @@ export function TrackForm({
   isSubmitting = false,
   id,
 }: TrackFormProps) {
-  const form = useForm<TrackFormData>({
-    resolver: zodResolver(TrackFormSchema),
-    defaultValues: {
-      title: initialData.title || "",
-      artist: initialData.artist || "",
-      album: initialData.album || "",
-      genres: initialData.genres || [],
-      coverImage: initialData.coverImage || "",
-    },
-    mode: "onChange",
-  });
-
-  const [selectedGenres, setSelectedGenres] = useState<string[]>(
-    initialData.genres || []
-  );
-  const [imageStatus, setImageStatus] = useState<
-    "idle" | "verifying" | "verified" | "error"
-  >("idle");
-  const [showPreview, setShowPreview] = useState(false);
-
   const {
-    data: availableGenres = [],
-    isLoading: isLoadingGenres,
-    isError: isErrorGenres,
-    error: genresError,
-    refetch: refetchGenres,
-  } = useGenres();
-
-  const coverImageUrl = form.watch("coverImage");
-
-  useEffect(() => {
-    const hasInitialImage =
-      !!initialData?.coverImage && initialData.coverImage === coverImageUrl;
-    if (hasInitialImage) {
-      setImageStatus("verified");
-      setShowPreview(true);
-    }
-  }, [initialData?.coverImage, coverImageUrl]);
-
-  const verifyImageUrl = useCallback(async (url: string): Promise<boolean> => {
-    if (!url) return false;
-    return new Promise((resolve) => {
-      const img = new window.Image();
-      const timeoutId = setTimeout(() => {
-        img.onload = null;
-        img.onerror = null;
-        img.src = "";
-        resolve(false);
-      }, 5000);
-      img.onload = () => {
-        clearTimeout(timeoutId);
-        resolve(true);
-      };
-      img.onerror = () => {
-        clearTimeout(timeoutId);
-        resolve(false);
-      };
-      img.src = url.includes("?")
-        ? `${url}&_cb=${Date.now()}`
-        : `${url}?cb=${Date.now()}`;
-    });
-  }, []);
-
-  const handleVerifyImage = useCallback(async () => {
-    const url = form.getValues("coverImage");
-    if (!url || imageStatus === "verifying") return;
-
-    setImageStatus("verifying");
-    setShowPreview(false);
-
-    const isValid = await verifyImageUrl(url);
-
-    if (isValid) {
-      setImageStatus("verified");
-      setShowPreview(true);
-      form.clearErrors("coverImage");
-    } else {
-      setImageStatus("error");
-      setShowPreview(false);
-      form.setError("coverImage", {
-        type: "manual",
-        message: "Unable to load image from this URL.",
-      });
-    }
-  }, [form, verifyImageUrl, imageStatus]);
-
-  useEffect(() => {
-    if (!coverImageUrl || imageStatus !== "idle") return;
-    const timer = setTimeout(() => {
-      const imageUrlPattern = /\.(jpeg|jpg|gif|png|webp|svg|avif)(\?.*)?$/i;
-      if (imageUrlPattern.test(coverImageUrl)) {
-        handleVerifyImage();
-      }
-    }, 800);
-    return () => clearTimeout(timer);
-  }, [coverImageUrl, handleVerifyImage, imageStatus]);
-
-  const handleClearImage = () => {
-    form.setValue("coverImage", "", { shouldValidate: true });
-    setShowPreview(false);
-    setImageStatus("idle");
-    form.clearErrors("coverImage");
-  };
-
-  const filteredGenres = availableGenres.filter(
-    (g) => !selectedGenres.includes(g)
-  );
-
-  const toggleGenre = (genre: string) => {
-    const newGenres = selectedGenres.includes(genre)
-      ? selectedGenres.filter((g) => g !== genre)
-      : [...selectedGenres, genre];
-    setSelectedGenres(newGenres);
-    form.setValue("genres", newGenres, { shouldValidate: true });
-  };
-
-  const handleRemoveGenre = (genre: string) => {
-    const newGenres = selectedGenres.filter((g) => g !== genre);
-    setSelectedGenres(newGenres);
-    form.setValue("genres", newGenres, { shouldValidate: true });
-  };
+    form,
+    selectedGenres,
+    imageStatus,
+    showPreview,
+    coverImageUrl,
+    isLoadingGenres,
+    isErrorGenres,
+    genresError,
+    filteredGenres,
+    handleVerifyImage,
+    handleClearImage,
+    toggleGenre,
+    handleRemoveGenre,
+    handleFormSubmit,
+    refetchGenres,
+  } = useTrackForm({ initialData, onSubmit, isSubmitting });
 
   return (
     <Form {...form}>
       <form
         id={id}
         data-testid="track-form"
-        onSubmit={form.handleSubmit(onSubmit)}
+        onSubmit={handleFormSubmit}
         className="space-y-6"
       >
         <div className="grid grid-cols-1 md:grid-cols-2 gap-6 p-1">
@@ -225,7 +118,6 @@ export function TrackForm({
               )}
             />
           </div>
-
           <div className="space-y-2">
             <FormField
               control={form.control}
@@ -244,8 +136,7 @@ export function TrackForm({
                         data-testid="input-cover-image"
                         onChange={(e) => {
                           field.onChange(e);
-                          setImageStatus("idle");
-                          setShowPreview(false);
+                          // Note: Image status reset logic is handled in the hook
                         }}
                       />
                     </FormControl>
@@ -314,12 +205,7 @@ export function TrackForm({
                             loading="lazy"
                             className="object-cover transition-all"
                             onError={() => {
-                              setImageStatus("error");
-                              setShowPreview(false);
-                              form.setError("coverImage", {
-                                type: "manual",
-                                message: "Failed to load image preview.",
-                              });
+                              // Error handling is managed in the hook
                             }}
                           />
                         </div>
@@ -356,7 +242,6 @@ export function TrackForm({
             />
           </div>
         </div>
-
         <FormField
           control={form.control}
           name="genres"
